Task_8_Snowflake/requirements.txt
snowflake-connector-python==3.0.2
apache-airflow-providers-snowflake==4.0.4
